A/B selection line were missing
1) Moved power lines from the mounting holes.
2) Added some ground vias.
3) Found that the !A/B selection lines were not routed to the TTL mux:
fixed.
4) moved ground planes to better shield the digital lines going to the
DAC
5) added ground plane on the bottom.

Remaining problem: the +-12V  lines still run inside the digital ground
section, therefore they capacitively couple with that potentially
"noisy" ground.
When connecting to USB the teensy can potentially inject the USB bus
interference  in the whole board: we could think about a future version
where we replicate the teensy directly on the PCB or we make our own
teensy with a USB isolator.
at this time we could just rely on an external USB opto- or
galvanic-isolator.